pyproject.toml
[project]
name = "bilibli-comment-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "bilibili-api-python>=17.4.0",
    "fastapi>=0.125.0",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
]
